feat(retryableWrites): adding more conditions for retries
Fixes NODE-1456
