+++ lib/sessions.js
@@ -6,7 +6,7 @@ const BSON = retrieveBSON();
 const Binary = BSON.Binary;
 const uuidV4 = require('./utils').uuidV4;
 const MongoError = require('./error').MongoError;
-const MongoNetworkError = require('./error').MongoNetworkError;
+const isRetryableError = require('./isRetryableError');
 
 function assertAlive(session, callback) {
   if (session.serverSession == null) {
@@ -217,35 +217,6 @@ class ClientSession extends EventEmitter {
       );
     });
   }
-}
-
-// see: https://github.com/mongodb/specifications/blob/master/source/retryable-writes/retryable-writes.rst#terms
-const RETRYABLE_ERROR_CODES = new Set([
-  6, // HostUnreachable
-  7, // HostNotFound
-  64, // WriteConcernFailed
-  89, // NetworkTimeout
-  91, // ShutdownInProgress
-  189, // PrimarySteppedDown
-  9001, // SocketException
-  11600, // InterruptedAtShutdown
-  11602, // InterruptedDueToReplStateChange
-  10107, // NotMaster
-  13435, // NotMasterNoSlaveOk
-  13436 // NotMasterOrSecondary
-]);
-
-function isRetryableError(error) {
-  if (
-    RETRYABLE_ERROR_CODES.has(error.code) ||
-    error instanceof MongoNetworkError ||
-    error.message.match(/not master/) ||
-    error.message.match(/node is recovering/)
-  ) {
-    return true;
-  }
-
-  return false;
 }
 
 function resetTransactionState(clientSession) {
