+++ lib/topologies/mongos.js
@@ -7,7 +7,6 @@ const BasicCursor = require('../cursor');
 const Logger = require('../connection/logger');
 const retrieveBSON = require('../connection/utils').retrieveBSON;
 const MongoError = require('../error').MongoError;
-const errors = require('../error');
 const Server = require('./server');
 const clone = require('./shared').clone;
 const diff = require('./shared').diff;
@@ -16,6 +15,7 @@ const createClientInfo = require('./shared').createClientInfo;
 const SessionMixins = require('./shared').SessionMixins;
 const isRetryableWritesSupported = require('./shared').isRetryableWritesSupported;
 const relayEvents = require('./shared').relayEvents;
+const isRetryableError = require('../isRetryableError');
 const BSON = retrieveBSON();
 
 /**
@@ -900,7 +900,7 @@ var executeWriteOperation = function(self, op, ns, ops, options, callback) {
 
   server[op](ns, ops, options, (err, result) => {
     if (!err) return callback(null, result);
-    if (!(err instanceof errors.MongoNetworkError) && !err.message.match(/not master/)) {
+    if (!isRetryableError(err)) {
       return callback(err);
     }
