+++ lib/topologies/replset.js
@@ -8,7 +8,6 @@ const BasicCursor = require('../cursor');
 const retrieveBSON = require('../connection/utils').retrieveBSON;
 const Logger = require('../connection/logger');
 const MongoError = require('../error').MongoError;
-const errors = require('../error');
 const Server = require('./server');
 const ReplSetState = require('./replset_state');
 const clone = require('./shared').clone;
@@ -18,6 +17,7 @@ const createClientInfo = require('./shared').createClientInfo;
 const SessionMixins = require('./shared').SessionMixins;
 const isRetryableWritesSupported = require('./shared').isRetryableWritesSupported;
 const relayEvents = require('./shared').relayEvents;
+const isRetryableError = require('../isRetryableError');
 
 const defaultAuthProviders = require('../auth/defaultAuthProviders').defaultAuthProviders;
 
@@ -1206,7 +1206,7 @@ function executeWriteOperation(args, options, callback) {
 
   const handler = (err, result) => {
     if (!err) return callback(null, result);
-    if (!(err instanceof errors.MongoNetworkError) && !err.message.match(/not master/)) {
+    if (!isRetryableError(err)) {
       return callback(err);
     }
